requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.6
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-loganalytics==12.0.0
azure-mgmt-securityinsight==1.0.0
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
cpe==1.2.1
cryptography==41.0.7
cybox==2.1.0.21
deepdiff==6.7.1
docopt==0.6.2
drawsvg==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastjsonschema==2.19.0
fqdn==1.5.1
frozenlist==1.4.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==5.13.0
isodate==0.6.1
isoduration==20.11.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter_core==5.6.0
loguru==0.7.2
lxml==4.9.4
maec==4.1.0.17
Markdown==3.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
mitreattack-python==2.0.14
mixbox==1.0.5
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.4
nbformat==5.9.2
netaddr==0.9.0
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
pluralizer==1.2.0
pooch==1.8.0
portalocker==2.8.2
pycountry==23.12.11
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.7
stix2-patterns==2.0.0
stix2-validator==3.1.4
stixmarx==1.0.8
tabulate==0.9.0
taxii2-client==2.3.0
tqdm==4.66.1
traitlets==5.14.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.1.0
weakrefmethod==1.0.3
webcolors==1.13
XlsxWriter==3.1.9
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
